requirements.txt
Flask
pdf2docx
gunicorn
flask-cors
reportlab
pypandoc
python-docx
Pillow